export_data_cf/terraform/src/requirements.txt
google-cloud-bigquery==3.11.0
google-cloud-storage==2.9.0